# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "bile-mo API",
        "description": "bile-mo is a web service offering a selection of mobile phones",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/bile-mo-api/v1",
            "description": "bile-mo API main server"
        }
    ],
    "paths": {
        "/admins": {
            "get": {
                "tags": [
                    "Admins management"
                ],
                "summary": "Returns list of admins",
                "operationId": "3ebdb43389a92d44e8d3701cb89cb43e",
                "responses": {
                    "200": {
                        "description": "HTTP_OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Ok",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Admins management"
                ],
                "summary": "Create a new admin",
                "operationId": "3ce12be3a00154c9e88bdc2993251ae4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {},
                                    "password": {},
                                    "roles": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "HTTP_CREATED",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Created",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "HTTP_CONFLICT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Conflict",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "HTTP_BAD_REQUEST",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Bad request",
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admins/{admin_id}": {
            "get": {
                "tags": [
                    "Admins management"
                ],
                "summary": "Returns admin by id",
                "operationId": "da91355e3bad24a471b2ffbe35c3a797",
                "parameters": [
                    {
                        "name": "admin_id",
                        "in": "path",
                        "description": "admin ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "HTTP_OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Ok",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserDTO"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "HTTP_NOT_FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Not found",
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Admins management"
                ],
                "summary": "Updates an admin by id",
                "operationId": "2e77786db77c13a28f8da099d6b175c4",
                "parameters": [
                    {
                        "name": "admin_id",
                        "in": "path",
                        "description": "admin ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {},
                                    "password": {},
                                    "roles": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "HTTP_CREATED",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Created",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserDTO"
                                    }
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "HTTP_CONFLICT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Conflict",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "HTTP_BAD_REQUEST",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Bad request",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "HTTP_NOT_FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Not found",
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admins management"
                ],
                "summary": "Deletes an admin by id",
                "operationId": "ccd4e6eb3bb68609b4de9823996d14bd",
                "parameters": [
                    {
                        "name": "admin_id",
                        "in": "path",
                        "description": "admin ID",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "HTTP_NO_CONTENT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "No content",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "HTTP_NOT_FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Not found",
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UserDTO": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "email": {
                        "type": "string",
                        "nullable": false
                    },
                    "password": {
                        "type": "string",
                        "nullable": false
                    },
                    "customer": {
                        "type": "array"
                    },
                    "roles": {
                        "type": "array",
                        "nullable": false
                    }
                },
                "type": "object"
            }
        }
    }
}